# Conflicts:
#	pom.xml
#	src/main/java/kz/logistic/pl/controllers/CustomerController.java
#	src/main/java/kz/logistic/pl/models/entities/CustomerEntity.java
#	src/main/java/kz/logistic/pl/models/entities/LoginEntity.java
#	src/main/java/kz/logistic/pl/models/entities/WarehouseEntity.java
#	src/main/java/kz/logistic/pl/models/pojos/Customer.java
#	src/main/java/kz/logistic/pl/models/pojos/impl/DefaultCustomer.java
#	src/main/java/kz/logistic/pl/repositories/AddressRepository.java
#	src/main/java/kz/logistic/pl/repositories/PaymentRepository.java
#	src/main/java/kz/logistic/pl/services/impl/DefaultCustomerService.java

+++ src/main/resources/db/migration/V1.0.1__create_table_login.sql
@@ -4,8 +4,8 @@ CREATE TABLE Login(
 			primary key,
 	username varchar(256),
 	password varchar(256),
-	roles_id integer,
-	customer_id integer,
+	roles_id bigserial,
+	customer_id bigserial,
 	seller_company_id bigserial,
-	shipper_id integer
+	shipper_id bigserial
 );
