+++ src/main/resources/db/migration/V1.0.9__create_table_warehouse.sql
@@ -5,7 +5,7 @@ CREATE TABLE Warehouse(
 	wh_name_kk varchar(256),
 	wh_name_ru varchar(256),
 	wh_name_en varchar(256),
-  wh_address_id integer,
-  seller_company_id integer,
+  wh_address_id bigserial,
+  seller_company_id bigserial,
   add_info varchar(500)
 );
