+++ src/main/resources/db/migration/V1.1.0__create_table_orders.sql
@@ -3,12 +3,12 @@ CREATE TABLE Orders(
 		constraint orders_pkey
 			primary key,
       order_number integer,
-      product_id integer,
-      seller_company_id integer,
+      product_id bigserial,
+      seller_company_id bigserial,
       order_date date,
       product_count integer,
       unit_price integer,
       total_price integer,
-      customer_id integer,
+      customer_id bigserial,
       delivering_status varchar(100)
 );
