+++ src/main/resources/db/migration/V1.1.1__create_table_shipping.sql
@@ -2,8 +2,8 @@ CREATE TABLE Shipping(
 	shipping_id bigserial not null
 		constraint shipping_pkey
 			primary key,
-      shipper_id integer,
-      order_id integer,
+      shipper_id bigserial,
+      order_id bigserial,
       order_number integer,
       shipping_address varchar(255),
       accept_date date,
